fix: Use Confd version and code from Makefile
The code repository and version from the Makefile was unused and was setting
to an outdated version. The template had the confd source repository and version
number hard coded.

I have fixed the template so that a user can define the confd source repository
and version number through the Makefile, the same way as for other dependencies.

Signed-off-by: Ronny Trommer <ronny@no42.org>